Move RolloverConfiguration to a more appropriate package
