+++ server/src/test/java/org/elasticsearch/action/admin/indices/rollover/RolloverConditionsTests.java
@@ -6,7 +6,7 @@
  * Side Public License, v 1.
  */
 
-package org.elasticsearch.cluster.metadata;
+package org.elasticsearch.action.admin.indices.rollover;
 
 import org.elasticsearch.TransportVersion;
 import org.elasticsearch.common.io.stream.Writeable;
